Add mip-map level generation

+++ convert_to_ktx.sh
@@ -3,7 +3,9 @@ cd output
 FILES=./*
 for f in $FILES
 do
-  echo "Processing $f file..."
-  EtcTool.exe ./$f -format RGB8A1 -effort 100 -output ./"${f%.*}.ktx"
-  rm $f
+  if [[ $f != *".ktx"* ]]; then
+    echo "Processing $f file..."
+    EtcTool.exe ./$f -format RGB8A1 -m 7 -effort 100 -output ./"${f%.*}.ktx"
+    rm $f
+  fi
 done
